Move resultsFormatter up into Users.js
