+++ lib/SearchAndSort/SearchAndSort.js
@@ -20,11 +20,11 @@ import transitionToParams from '@folio/stripes-components/util/transitionToParam
 import Notes from '@folio/stripes-smart-components/lib/Notes';
 import removeQueryParam from '../../removeQueryParam';
 import contactTypes from '../../data/contactTypes';
-import { getFullName } from '../../util';
 
 
 class SearchAndSort extends React.Component {
   static propTypes = {
+    resultsFormatter: PropTypes.shape({}),
     translationBase: PropTypes.string.isRequired,
     viewRecordComponent: PropTypes.func.isRequired,
     editRecordComponent: PropTypes.func.isRequired, // eslint-disable-line react/no-unused-prop-types
@@ -297,7 +297,7 @@ class SearchAndSort extends React.Component {
       placeholder={stripes.intl.formatMessage({ id: `${translationBase}.search` })}
     />);
 
-    const newUserButton = (
+    const newRecordButton = (
       <IfPermission perm="users.item.post">
         <IfPermission perm="login.item.post">
           <IfPermission perm="perms.users.item.post">
@@ -308,18 +308,6 @@ class SearchAndSort extends React.Component {
         </IfPermission>
       </IfPermission>
     );
-
-    const resultsFormatter = {
-      Status: user => (user.active ? 'Active' : 'Inactive'),
-      Name: user => getFullName(user),
-      Barcode: user => user.barcode,
-      'Patron Group': (user) => {
-        const pg = patronGroups.filter(g => g.id === user.patronGroup)[0];
-        return pg ? pg.group : '?';
-      },
-      Username: user => user.username,
-      Email: user => _.get(user, ['personal', 'email']),
-    };
 
     const detailsPane = (
       stripes.hasPerm('users.item.get') ?
@@ -364,7 +352,7 @@ class SearchAndSort extends React.Component {
               </div>
             </div>
           }
-          lastMenu={!this.props.disableRecordCreation ? newUserButton : null}
+          lastMenu={!this.props.disableRecordCreation ? newRecordButton : null}
           noOverflow
         >
           <MultiColumnList
@@ -372,7 +360,7 @@ class SearchAndSort extends React.Component {
             contentData={records}
             selectedRow={this.state.selectedItem}
             rowMetadata={['id', 'username']}
-            formatter={resultsFormatter}
+            formatter={this.props.resultsFormatter}
             onRowClick={this.onSelectRow}
             onHeaderClick={this.onSort}
             onNeedMoreData={this.onNeedMore}
